pyproject.toml
[project]
name = "llmops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "google-cloud-aiplatform>=1.120.0",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-storage>=2.19.0",
    "kfp>=2.7.0",
    "pandas>=2.2.2",
    "datasets>=3.0.1",
    "gcsfs>=2024.6.1",
    "pyarrow>=17.0.0",
    "python-dotenv>=1.0.1",
    "ipykernel>=7.0.1",
    "transformers==4.46.*",
    "peft==0.13.2",
    "accelerate>=0.33",
    "tensorboard>=2.17",
    "scipy>=1.11",
]
